+++ src/main/java/tornadofx/Properties.kt
@@ -1,7 +1,6 @@
 package tornadofx
 
 import javafx.beans.property.*
-import javafx.beans.value.ObservableValue
 import java.lang.reflect.Field
 import kotlin.properties.ReadWriteProperty
 import kotlin.reflect.*
@@ -180,17 +179,6 @@ private class UnsynchronizedSingleAssign<T> : SingleAssign<T> {
 
     override fun isInitialized() = initialized
 }
-
-fun <T> Property<T>.getObservableValue(): ObservableValue<T>? {
-    val field = javaClass.findFieldByName("observable")
-    if (field != null) {
-        field.isAccessible = true
-        val observableValue = field.get(this)
-        if (observableValue is ObservableValue<*>) return observableValue as ObservableValue<T>
-    }
-    return null
-}
-
 operator fun <T> Property<T>.getValue(thisRef: Any?, property: KProperty<*>): T = value
 operator fun <T> Property<T>.setValue(thisRef: Any?, property: KProperty<*>, value: T) {
     this.value = value