+++ CHANGELOG.md
@@ -12,13 +12,14 @@ All notable changes to this project will be documented in this file.
 - `runAsync` on Injectable class references (`CustomerController::class.runAsync { listContacts(customerId) }`)
 - `runAsync` on Injectable class function references (`CustomerController::listContacts.runAsync(customerId)`)
 - `ObservableValue.onChange` listener
+- `UIComponent.addOnDockListener` and `UIComponent.addOnUndockListener`
 - LayoutDebugger (https://github.com/edvin/tornadofx/wiki/Layout-Debugger)
 - ViewModel (https://github.com/edvin/tornadofx/wiki/ViewModel)
 - TableView `cellDecorator` 
 - ComboBox `cellFormat` formatter function
 - TreeView `lazyPopulate` alternative to `populate` that lazily creates children as needed
 - TreeItem nesting extension functions (https://github.com/edvin/tornadofx/issues/134)
-- TableView `selectWhere()`, `moveToTopWhere()` and `BottomWhere()` (https://github.com/edvin/tornadofx/issues/134)
+- TableView `selectWhere()`, `moveToTopWhere()` and `moveToBottomWhere()` (https://github.com/edvin/tornadofx/issues/134)
 - Group builder `group`
 
 ### Changed